+++ src/ext_swapchain_colorspace.rs
@@ -14,6 +14,6 @@
 
 //! [`VK_EXT_swapchain_colorspace`](https://www.khronos.org/registry/vulkan/specs/1.0-extensions/html/vkspec.html#VK_EXT_swapchain_colorspace)
 
-pub const VK_EXT_SWAPCHAIN_COLOR_SPACE_SPEC_VERSION: u32 = 2;
+pub const VK_EXT_SWAPCHAIN_COLOR_SPACE_SPEC_VERSION: u32 = 3;
 pub const VK_EXT_SWAPCHAIN_COLOR_SPACE_EXTENSION_NAME: &'static [u8; 28] = b"VK_EXT_swapchain_colorspace\x00";
 pub const VK_EXT_SWAPCHAIN_COLOR_SPACE_EXTENSION_NAME_STR: &'static str = "VK_EXT_swapchain_colorspace";